// packages/frontend/src/openapi-services/cyclop/models/api-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cyclop APIs
 * Welcome to the Cyclop API service\'s documentation. You can play around with the different API endpoints to familiarise yourself with. --- **Release notes:**   * Improvments in CI/CD   * Platform endpoint   * Update password endpoint   * Profiles Endpoint   * JWT token revogation    * Test coverage improvment 10% - 20%   * Unit tests fix 
 *
 * The version of the OpenAPI document: {version}
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BSD } from './bsd';

/**
 * 
 * @export
 * @interface ApiResponse
 */
export interface ApiResponse {
    /**
     * 
     * @type {string}
     * @memberof ApiResponse
     */
    'status'?: ApiResponseStatusEnum;
    /**
     * 
     * @type {Array<BSD>}
     * @memberof ApiResponse
     */
    'data'?: Array<BSD>;
    /**
     * 
     * @type {string}
     * @memberof ApiResponse
     */
    'message'?: string;
}

export const ApiResponseStatusEnum = {
    Success: 'success',
    Fail: 'fail'
} as const;

export type ApiResponseStatusEnum = typeof ApiResponseStatusEnum[keyof typeof ApiResponseStatusEnum];